share getState between server and client
