+++ lib/app-manager.js
@@ -5,6 +5,7 @@ import { setNames, retry, defaultGetAppNameFromUrl, getErrorStatusDetails, getLo
 import shimHistory from './utils/shim-history';
 import initReconcile from './utils/reconcile';
 import initLifecycle from './utils/lifecycle';
+import initGetState from './utils/get-state';
 
 import type {
   EventsType,
@@ -14,7 +15,6 @@ import type {
   StateChangeType,
   SlotNameType,
   FragmentNameType,
-  AppType,
 } from './index';
 
 import type Window from './utils/window-stub';
@@ -37,6 +37,10 @@ export default function initAppManager(w: Window) {
 
   function selectFromDom(querySelector: string): ?Element {
     return w.document.querySelector(querySelector);
+  }
+
+  function getBrowserLocation(): { pathname: string, search: string } {
+    return w.location;
   }
 
   function appManager(config: ConfigType, events: EventsType, overrideOptions?: OverrideOptionsType) {
@@ -118,66 +122,14 @@ export default function initAppManager(w: Window) {
 
     const reconcile = initReconcile(config);
     const lifecycle = initLifecycle(w, fragments, getSlotElement, options);
-
-    async function getState(prevApp: ?AppType, eventTitle?: mixed): Promise<?StateType> {
-      const { pathname, search } = w.location;
-
-      const url = `${pathname}${search}`;
-
-      const errData = {
-        source: 'set_state',
-        fragment: null,
-        slot: null,
-        path: pathname,
-      };
-
-      let appName = null;
-
-      if (typeof options.getAppNameFromUrl === 'function') {
-        try {
-          appName = await options.getAppNameFromUrl(url);
-        } catch (err) {
-          throw new AppManagerError(err, Object.assign({}, errData, {
-            app: appName,
-            level: levels.ERROR,
-            code: 'get_app_name_from_location',
-            recoverable: false,
-          }));
-        }
-      }
-
-      if (!appName) {
-        return null;
-      }
-
-      let additionalState = {};
-
-      if (typeof options.getAdditionalState === 'function') {
-        try {
-          additionalState = await options.getAdditionalState(appName, url);
-        } catch (err) {
-          throw new AppManagerError(err, Object.assign({}, errData, {
-            app: appName,
-            level: levels.ERROR,
-            code: 'get_additional_state',
-            recoverable: false,
-          }));
-        }
-      }
-
-      return Object.assign({}, additionalState, {
-        event: eventTitle && typeof eventTitle === 'string' ? eventTitle : null,
-        path: pathname,
-        prevApp,
-        app: apps[appName],
-      });
-    }
+    const getState = initGetState(apps, options);
 
     function getStateChanger(initialRender: boolean, eventTitle?: mixed) {
       return async function stateChanger(): Promise<void> {
         try {
           const currentApp = state ? state.app : null;
-          const newState = await getState(currentApp, eventTitle);
+          const { pathname, search } = getBrowserLocation();
+          const newState = await getState(`${pathname}${search}`, currentApp, eventTitle);
 
           if (!newState) {
             events.emit(eventTitles.EXTERNAL_LINK);
