+++ lib/server.js
@@ -6,6 +6,7 @@ import AppManagerError from './utils/app-manager-error';
 import { setNames, defaultGetAppNameFromUrl } from './utils/config';
 import { levels } from './constants';
 import initReconcile from './utils/reconcile';
+import initGetState from './utils/get-state';
 
 export default function appManager(config: ConfigType, overrideOptions?: OptionsType) {
   const apps = setNames(config.apps);
@@ -17,6 +18,7 @@ export default function appManager(config: ConfigType, overrideOptions?: Options
 
   const options = Object.assign({}, defaultOptions, overrideOptions || {});
   const reconcile = initReconcile(config);
+  const getState = initGetState(apps, options);
 
   return {
     async getSlotsMarkup(url: string, ...getMarkupArgs: Array<mixed>): Promise<{ [slotName: SlotNameType]: string }> {
@@ -50,26 +52,15 @@ export default function appManager(config: ConfigType, overrideOptions?: Options
         }));
       }
 
-      let additionalState = {};
+      const state = await getState(url);
 
-      if (typeof options.getAdditionalState === 'function') {
-        try {
-          additionalState = await options.getAdditionalState(currentAppName);
-        } catch (err) {
-          throw new AppManagerError(`App ${currentAppName} does not have a fragments array`, Object.assign({}, errData, {
-            level: levels.ERROR,
-            code: 'get_additional_state',
-            recoverable: false,
-          }));
-        }
+      if (!state) {
+        throw new AppManagerError('Could not get app state', Object.assign({}, errData, {
+          level: levels.ERROR,
+          code: 'get_state',
+          recoverable: false,
+        }));
       }
-
-      const state = Object.assign({}, additionalState, {
-        event: null,
-        path,
-        prevApp: null,
-        app: apps[currentAppName],
-      });
 
       const { mount } = reconcile(state, true);
