+++ lib/utils/path.js
@@ -7,7 +7,7 @@ import type { AppNameType } from '../index';
 
 type ParamsType = { [paramName: string]: string };
 
-export default function getPathHelpers(appPaths: { [appName: AppNameType]: string }) {
+export default function getPathHelpers(appPaths: { [appName: AppNameType]: { appPath?: string, appPaths?: Array<string> } }) {
   const pathParams = Object.keys(appPaths || {}).reduce((evaluatedPaths, appName) => {
     const params = pathToRegexp.parse(appPaths[appName])
       .filter((key) => typeof key === 'object')
